% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_in_network.R
\name{address_in_network}
\alias{address_in_network}
\alias{is_within}
\alias{is_within_any}
\title{Are IP addresses within IP network(s)?}
\usage{
is_within(address, network)

is_within_any(address, network)
}
\arguments{
\item{address}{An \code{ip_address} vector}

\item{network}{An \code{ip_network} vector}
}
\description{
\code{is_within()} checks if an IP address is within a single IP network

\code{is_within_any()} checks if an IP address is within a vector of IP networks
}
\examples{
address <- ip_address(c("192.168.0.1", "2001:db8::8a2e:370:7334", "1.2.3.4"))
network1 <- ip_network(c("192.168.0.0/16", "2001:db8::/36", "0.0.0.0/30"))
network2 <- ip_network(c("2001:db8::/36", "192.168.0.0/16", "0.0.0.0/30"))

is_within(address, network1)
is_within(address, network2)

is_within_any(address, network1)
is_within_any(address, network2)

}
