% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc_last_ip_dec}
\alias{calc_last_ip_dec}
\title{Calculate hosts and last ip from cidr}
\usage{
calc_last_ip_dec(ip_dec_vect, prefix_vect)
}
\arguments{
\item{ip_dec_vect}{Vector of decimal IP addresses}

\item{prefix_vect}{Vector of cidr network prefixes}
}
\value{
last_ip_dec_vector Vector of decimal last ips
}
\description{
Calculate hosts and last ip from cidr
}
