% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_version.R
\name{classify_version}
\alias{classify_version}
\alias{is_ipv4}
\alias{is_ipv6}
\title{Classify version of IP address space}
\usage{
is_ipv4(x)

is_ipv6(x)
}
\arguments{
\item{x}{An \code{ip_address} or \code{ip_network} vector}
}
\description{
\code{is_ipv4} identifies IPv4 addresses/networks

\code{is_ipv6} identifies IPv6 addresses/networks
}
\examples{
tibble(ip = ip_address(c("0.0.0.1", "192.168.0.1", "2001:db8::7334"))) \%>\%
  mutate(ipv4 = is_ipv4(ip), ipv6 = is_ipv6(ip))

}
