% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_ranges.R
\name{collapse_networks}
\alias{collapse_networks}
\title{Collapse contiguous and overlapping networks}
\usage{
collapse_networks(network)
}
\arguments{
\item{network}{An \code{\link{ip_network}} vector}
}
\value{
An \code{\link{ip_network}} vector (potentially shorter than the input)
}
\description{
Given a vector of networks, this returns the minimal set of networks required
to represent the same range of addresses.
}
\examples{
collapse_networks(ip_network(c("192.168.0.0/24", "192.168.1.0/24")))
}
\seealso{
\code{\link[=exclude_networks]{exclude_networks()}}
}
