% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_network.R
\name{common_network}
\alias{common_network}
\title{Find the common network of two addresses}
\usage{
common_network(address1, address2)
}
\arguments{
\item{address1}{An \code{\link{ip_address}} vector}

\item{address2}{An \code{\link{ip_address}} vector}
}
\value{
An \code{\link{ip_network}} vector
}
\description{
Returns the smallest network that contains both addresses.

This is a simple way to construct a network from its first and last addresses.
However, if the addresses do not correspond to network boundaries, then the
returned network will extend beyond their address range.
}
\examples{
first <- ip_address(c("192.168.0.0", "224.0.0.0"))
last <- ip_address(c("192.168.0.15", "239.255.255.255"))

common_network(first, last)
}
