% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dec2dot}
\alias{dec2dot}
\title{Convert decimal ip to dot decimal string}
\usage{
dec2dot(ip_dec_vect)
}
\arguments{
\item{ip_dec_vect}{Decimal IP to convert}
}
\value{
ip_dot_vect Dot decimal string
}
\description{
Convert decimal ip to dot decimal string
}
