% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hostname.R
\name{hostname}
\alias{hostname}
\alias{as_hostname}
\alias{from_hostname}
\title{Represent address as hostname}
\usage{
as_hostname(ip, simplify = TRUE)

from_hostname(host, simplify = TRUE)
}
\arguments{
\item{ip}{An \code{\link{ip_address}} vector}

\item{host}{A character vector of hostnames}
}
\description{
Represent address as hostname
}
\seealso{
\code{\link[=nsl]{nsl()}}
}
