% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iana_ipv4}
\alias{iana_ipv4}
\title{IPv4 address space allocation}
\format{
A data frame with 122 rows and 3 variables:
\describe{
\item{network}{Address block (an \code{\link{ip_network}} vector)}
\item{allocation}{There are three types of allocation:
\itemize{
\item reserved
\item managed by regional Internet registry (RIR)
\item assigned to organization
}
}
\item{label}{The RIR, organization or purpose for reservation}
}
}
\source{
\url{https://www.iana.org/assignments/ipv4-address-space}
}
\usage{
iana_ipv4
}
\description{
A dataset containing the registry of allocated blocks in IPv4 address space.
}
\note{
Last updated 2020-08-18
}
\examples{
iana_ipv4
}
\seealso{
\code{\link[=is_reserved]{is_reserved()}}
}
\keyword{datasets}
