% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iana_ipv6}
\alias{iana_ipv6}
\title{IPv6 address space allocation}
\format{
A data frame with 47 rows and 3 variables:
\describe{
\item{network}{Address block (an \code{\link{ip_network}} vector)}
\item{allocation}{There are two types of allocation:
\itemize{
\item reserved
\item managed by regional Internet registry (RIR)
}
}
\item{label}{The RIR or purpose for reservation}
}
}
\source{
\url{https://www.iana.org/assignments/ipv6-address-space}

\url{https://www.iana.org/assignments/ipv6-unicast-address-assignments}
}
\usage{
iana_ipv6
}
\description{
A dataset containing the registry of allocated blocks in IPv6 address space.
}
\note{
Last updated 2020-08-18
}
\examples{
iana_ipv6
}
\seealso{
\code{\link[=is_reserved]{is_reserved()}}
}
\keyword{datasets}
