% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip_interface.R
\name{ip_interface}
\alias{ip_interface}
\alias{ip_interface.default}
\alias{ip_interface.ip_address}
\alias{as_ip_interface}
\alias{is_ip_interface}
\alias{format.ip_interface}
\alias{as.character.ip_interface}
\title{Vector of IP interfaces}
\usage{
ip_interface(...)

\method{ip_interface}{default}(ip = character(), ...)

\method{ip_interface}{ip_address}(address, prefix_length, ...)

as_ip_interface(x)

is_ip_interface(x)

\method{format}{ip_interface}(x, ...)

\method{as.character}{ip_interface}(x, ...)
}
\arguments{
\item{...}{Arguments to be passed to other methods}

\item{ip}{Character vector of IP addresses, in dot-decimal notation (IPv4)
or hexadecimal notation (IPv6).}

\item{address}{An \code{\link{ip_address}} vector}

\item{prefix_length}{An integer vector}

\item{x}{An \code{ip_interface} vector}
}
\value{
An \code{ip_interface} vector
}
\description{
\code{ip_interface()} constructs a vector of IP interfaces.

\code{as_ip_interface()} casts an object to \code{ip_interface}.

\code{is_ip_interface()} checks if an object is of class \code{ip_interface}.
}
\seealso{
\code{\link[=ip_address]{ip_address()}}, \code{\link[=ip_network]{ip_network()}}
}
