% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip_address.R, R/ip_interface.R, R/ip_network.R
\name{ip_test}
\alias{ip_test}
\alias{is_ip_address}
\alias{is_ip_interface}
\alias{is_ip_network}
\title{Test for IP vector}
\usage{
is_ip_address(x)

is_ip_interface(x)

is_ip_network(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
A logical scalar
}
\description{
Check if an object is an \code{\link{ip_address}}, \code{\link{ip_network}} or \code{\link{ip_interface}}
vector.
}
\examples{
is_ip_address(ip_address("192.168.0.1"))
is_ip_interface(ip_interface("192.168.0.1/10"))
is_ip_network(ip_network("192.168.0.0/24"))
}
