% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_repr.R
\name{ip_to_bits}
\alias{ip_to_bits}
\alias{bits_to_ip}
\title{Represent address as sequence of bits}
\usage{
ip_to_bits(x)

bits_to_ip(x)
}
\arguments{
\item{x}{\itemize{
\item For \code{ip_to_bits()}: An \code{\link{ip_address}} vector
\item For \code{bits_to_ip()}: A character vector containing only \code{0} and \code{1} characters
}}
}
\value{
\itemize{
\item For \code{ip_to_bits()}: A character vector
\item For \code{bits_to_ip()}: An \code{\link{ip_address}} vector
}
}
\description{
Encode or decode an \code{\link{ip_address}} vector to a character vector of bits.
}
\details{
The bits are stored in network order (also known as big-endian order), which
is part of the IP standard.

IPv4 addresses use 32 bits, IPv6 addresses use 128 bits, and missing values
are encoded as \code{NA}.
}
\examples{
x <- ip_address(c("192.168.0.1", "2001:db8::8a2e:370:7334", NA))
ip_to_bits(x)

bits_to_ip(ip_to_bits(x))
}
\seealso{
\itemize{
\item \code{\link[=ip_to_integer]{ip_to_integer()}} and \code{\link[=integer_to_ip]{integer_to_ip()}}
\item \code{\link[=ip_to_bytes]{ip_to_bytes()}} and \code{\link[=bytes_to_ip]{bytes_to_ip()}}
}
}
