% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipaddress-package.R
\docType{package}
\name{ipaddress-package}
\alias{ipaddress}
\alias{ipaddress-package}
\title{ipaddress: Data Analysis for IP Addresses and Networks}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Classes and functions for working with IP (Internet Protocol) addresses and networks, inspired by the Python 'ipaddress' module. Offers full support for both IPv4 and IPv6 (Internet Protocol versions 4 and 6) address spaces. It is specifically designed to work well with the 'tidyverse'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://davidchall.github.io/ipaddress/}
  \item \url{https://github.com/davidchall/ipaddress}
  \item Report bugs at \url{https://github.com/davidchall/ipaddress/issues}
}

}
\author{
\strong{Maintainer}: David Hall \email{david.hall.physics@gmail.com} (\href{https://orcid.org/0000-0002-2193-0480}{ORCID})

Authors:
\itemize{
  \item David Hall \email{david.hall.physics@gmail.com} (\href{https://orcid.org/0000-0002-2193-0480}{ORCID})
}

}
\keyword{internal}
