% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipaddress-package.R, R/vctrs-arithmetic.R,
%   R/vctrs-cast.R, R/vctrs-coerce.R
\name{ipaddress-vctrs}
\alias{ipaddress-vctrs}
\alias{vec_arith.ip_address}
\alias{vec_arith.ip_interface}
\alias{vec_cast.ip_address}
\alias{vec_cast.ip_network}
\alias{vec_cast.ip_interface}
\alias{vec_ptype2.ip_address}
\alias{vec_ptype2.ip_network}
\alias{vec_ptype2.ip_interface}
\title{Internal vctrs methods}
\usage{
\method{vec_arith}{ip_address}(op, x, y, ...)

\method{vec_arith}{ip_interface}(op, x, y, ...)

\method{vec_cast}{ip_address}(x, to, ...)

\method{vec_cast}{ip_network}(x, to, ...)

\method{vec_cast}{ip_interface}(x, to, ...)

\method{vec_ptype2}{ip_address}(x, y, ...)

\method{vec_ptype2}{ip_network}(x, y, ...)

\method{vec_ptype2}{ip_interface}(x, y, ...)
}
\description{
Internal vctrs methods
}
\keyword{internal}
