% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipv6_transition.R
\name{ipv6_transition}
\alias{ipv6_transition}
\alias{is_ipv4_mapped}
\alias{extract_ipv4_mapped}
\alias{is_6to4}
\alias{extract_6to4}
\alias{is_teredo}
\alias{extract_teredo_server}
\alias{extract_teredo_client}
\title{IPv6 transition mechanisms}
\usage{
is_ipv4_mapped(x)

extract_ipv4_mapped(x)

is_6to4(x)

extract_6to4(x)

is_teredo(x)

extract_teredo_server(x)

extract_teredo_client(x)
}
\arguments{
\item{x}{An \code{\link{ip_address}} vector}
}
\value{
An \code{\link{ip_address}} vector
}
\description{
There are multiple mechanisms designed to help with the transition
from IPv4 to IPv6, which usually involve embedding an IPv4 address
within an IPv6 address. These functions make it possible to extract
the embedded IPv4 addresses.
}
\details{
Addresses already in the target address space are unaffected by
these functions.
}
\examples{
is_ipv4_mapped(ip_network("::ffff:0.0.0.0/96"))

extract_ipv4_mapped(ip_address("::ffff:192.0.2.128"))

is_6to4(ip_network("2002::/16"))

extract_6to4(ip_address("2002:c000:0204::"))

is_teredo(ip_network("2001::/32"))

extract_teredo_server(ip_address("2001:0000:4136:e378:8000:63bf:3fff:fdd2"))

extract_teredo_client(ip_address("2001:0000:4136:e378:8000:63bf:3fff:fdd2"))
}
