% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{netmask}
\alias{netmask}
\alias{hostmask}
\title{Netmask and hostmask}
\usage{
netmask(x)

hostmask(x)
}
\arguments{
\item{x}{An \code{ip_network} vector}
}
\value{
An \code{ip_address} vector corresponding to the netmask
}
\description{
\code{netmask()} extracts the netmask from an IP network

\code{hostmask()} extracts the hostmask from an IP network
}
\examples{
netmask(ip_network(c("192.168.1.0/24", "2001:db00::0/24")))

hostmask(ip_network(c("192.168.1.0/24", "2001:db00::0/24")))

}
