% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip_network.R
\name{network_info}
\alias{network_info}
\alias{prefix_length}
\alias{num_addresses}
\alias{network_address}
\alias{broadcast_address}
\title{Basic information about a network}
\usage{
prefix_length(x)

num_addresses(x)

network_address(x)

broadcast_address(x)
}
\arguments{
\item{x}{An \code{\link{ip_network}} vector}
}
\value{
\itemize{
\item \code{prefix_length()} returns an integer vector
\item \code{num_addresses()} returns a numeric vector
\item \code{network_address()} and \code{broadcast_address()} return an \code{\link{ip_address}} vector
}
}
\description{
\code{prefix_length()} gives the prefix length

\code{num_addresses()} gives the total number of addresses

\code{network_address()} gives the first address

\code{broadcast_address()} gives the last address
}
\details{
The broadcast address is a special address at which any host connected
to the network can receive messages. That is, packets sent to this address
are received by all hosts on the network.
In IPv4, the last address of a network is the broadcast address.
Although IPv6 does not follow this approach to broadcast addresses, the
\code{broadcast_address()} function still returns the last address of the network.
}
\examples{
x <- ip_network(c("192.168.0.0/22", "2001:db00::0/26"))

prefix_length(x)

num_addresses(x)

network_address(x)

broadcast_address(x)

}
\seealso{
The prefix length can equivalently be represented by the \code{\link[=netmask]{netmask()}} or \code{\link[=hostmask]{hostmask()}}.
}
