% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample_ip}
\alias{sample_ip}
\alias{sample_ip.default}
\alias{sample_ip.ip_network}
\title{Sample addresses within a network}
\usage{
sample_ip(...)

\method{sample_ip}{default}(space = c("IPv4", "IPv6"), size, replace = FALSE, ...)

\method{sample_ip}{ip_network}(x, size, replace = FALSE, ...)
}
\arguments{
\item{...}{Arguments to be passed to other methods}

\item{space}{The address space to sample ("IPv4" or "IPv6")}

\item{size}{Integer specifying the number of addresses to return}

\item{replace}{Should sampling be with replacement?}

\item{x}{An \code{\link{ip_network}} scalar}
}
\value{
An \code{\link{ip_address}} vector
}
\description{
Generates random addresses within an IP network or from the
entire address space.
}
\examples{
sample_ip("IPv4", 5)

sample_ip("IPv6", 5)

sample_ip(ip_network("192.168.0.0/16"), 5)

sample_ip(ip_network("2001:db8::/48"), 5)
}
\seealso{
Use \code{\link[=seq.ip_network]{seq.ip_network()}} to generate \emph{all} addresses in a network.
}
