% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample_network}
\alias{sample_network}
\title{Sample addresses within a network}
\usage{
sample_network(x, size, replace = FALSE)
}
\arguments{
\item{x}{An \code{\link{ip_network}} scalar}

\item{size}{Integer specifying the number of addresses to return}

\item{replace}{Should sampling be with replacement?}
}
\value{
An \code{\link{ip_address}} vector
}
\description{
Generates random addresses within an IP network.
}
\examples{
sample_network(ip_network("192.168.0.0/16"), 5)

sample_network(ip_network("2001:db8::/48"), 5)
}
\seealso{
Use \code{\link[=seq.ip_network]{seq.ip_network()}} to generate \emph{all} addresses in a network.

Use \code{\link[=sample_ipv4]{sample_ipv4()}} or \code{\link[=sample_ipv6]{sample_ipv6()}} to sample addresses from
the entire address space.
}
