% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-vctrs.R, R/ip_address.R, R/ip_network.R
\name{vctrs-compat}
\alias{vctrs-compat}
\alias{vec_cast.vctrs_ip_address}
\alias{vec_cast.vctrs_ip_address.character}
\alias{vec_ptype2.vctrs_ip_address}
\alias{vec_cast.vctrs_ip_network}
\alias{vec_cast.vctrs_ip_network.character}
\alias{vec_ptype2.vctrs_ip_network}
\title{vctrs compatibility functions}
\usage{
\method{vec_cast}{vctrs_ip_address}(x, to, ...)

\method{vec_cast.vctrs_ip_address}{character}(x, to, ...)

\method{vec_ptype2}{vctrs_ip_address}(x, y, ...)

\method{vec_cast}{vctrs_ip_network}(x, to, ...)

\method{vec_cast.vctrs_ip_network}{character}(x, to, ...)

\method{vec_ptype2}{vctrs_ip_network}(x, y, ...)
}
\arguments{
\item{x}{Objects.}

\item{to}{Type to cast to.}

\item{...}{Used to pass along error message information.}

\item{y}{Vector types.}
}
\description{
These functions allow ipaddress objects to work with vctrs.

Double dispatch methods to support \code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}}.

Double dispatch methods to support \code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}}.
}
\keyword{internal}
