% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-arithmetic.R
\name{vec_arith.ip_address}
\alias{vec_arith.ip_address}
\title{Arithmetic operations}
\usage{
\method{vec_arith}{ip_address}(op, x, y, ...)
}
\arguments{
\item{op}{An arithmetic operator as a string}

\item{x}{A pair of vectors. For \code{!}, unary \code{+} and unary \code{-}, \code{y} will be
a sentinel object of class \code{MISSING}, as created by \code{MISSING()}.}

\item{y}{A pair of vectors. For \code{!}, unary \code{+} and unary \code{-}, \code{y} will be
a sentinel object of class \code{MISSING}, as created by \code{MISSING()}.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Double dispatch methods to support \code{\link[vctrs:vec_arith]{vctrs::vec_arith()}}.
}
\keyword{internal}
