% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-cast.R
\name{vec_cast.ip_network}
\alias{vec_cast.ip_network}
\title{Casting}
\usage{
\method{vec_cast}{ip_network}(x, to, ...)
}
\arguments{
\item{x}{Vectors to cast.}

\item{to}{Type to cast to. If \code{NULL}, \code{x} will be returned as is.}

\item{...}{For \code{vec_cast_common()}, vectors to cast. For
\code{vec_cast()} and \code{vec_restore()}, these dots are only for future
extensions and should be empty.}
}
\value{
A vector the same length as \code{x} with the same type as \code{to},
or an error if the cast is not possible. An error is generated if
information is lost when casting between compatible types (i.e. when
there is no 1-to-1 mapping for a specific value).
}
\description{
Double dispatch methods to support \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}}.
}
\keyword{internal}
