% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-coerce.R
\name{vec_ptype2.ip_network}
\alias{vec_ptype2.ip_network}
\title{Coercion}
\usage{
\method{vec_ptype2}{ip_network}(x, y, ...)
}
\arguments{
\item{x}{Vector types.}

\item{y}{Vector types.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A prototype if \code{x} and \code{y} can be safely coerced to the same
prototype; otherwise it returns an error.
}
\description{
Double dispatch methods to support \code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}}.
}
\keyword{internal}
