% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within_network.R
\name{address_in_network}
\alias{address_in_network}
\alias{is_within}
\alias{is_within_any}
\title{Are IP addresses within IP network(s)?}
\usage{
is_within(address, network)

is_within_any(address, network)
}
\arguments{
\item{address}{An \code{ip_address} vector}

\item{network}{An \code{ip_network} vector}
}
\description{
\code{is_within()} checks if an IP address is within a single IP network

\code{is_within_any()} checks if an IP address is within a vector of IP networks
}
\examples{
is_within(ip_address("192.0.2.6"), ip_network("192.0.2.0/28"))
is_within(ip_address("192.0.3.6"), ip_network("192.0.2.0/28"))

is_within_any(ip_address("192.0.3.6"), ip_network(c("192.0.2.0/28", "192.0.3.0/28")))

}
