% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{binary}
\alias{binary}
\alias{as_binary}
\alias{from_binary}
\title{Encode or decode address as binary}
\usage{
as_binary(ip)

from_binary(bin)
}
\arguments{
\item{ip}{An \code{\link{ip_address}} vector}

\item{bin}{A character vector containing only \code{0} and \code{1} characters}
}
\description{
Encode or decode address as binary
}
\details{
The bits are stored in network order (also known as
big-endian order), which is part of the IP standard.

IPv4 addresses use 32 bits, IPv6 addresses use 128 bits,
and missing values are encoded as \code{NA}.
}
\examples{
x <- ip_address(c("192.168.0.1", "2001:db8::8a2e:370:7334", NA))
as_binary(x)

from_binary(as_binary(x))
}
