% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country.R
\name{country_networks}
\alias{country_networks}
\title{Country-level IP networks}
\source{
\url{https://www.iwik.org/ipcountry} (updated daily)
}
\usage{
country_networks(country)
}
\arguments{
\item{country}{Character vector of two-letter country codes (ISO 3166-1 alpha-2)}
}
\value{
A data frame with 2 variables:
\itemize{
\item \code{country}: A character vector
\item \code{network}: A list of \code{\link{ip_network}} vectors
}
}
\description{
Retrieve list of IP networks registered to specific countries.
}
\details{
This function requires an internet connection to download network lists.
}
\examples{
country_networks(c("BR", "FR", "IN", "NA", "US"))

}
