% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_exclude.R
\name{exclude_networks}
\alias{exclude_networks}
\title{Remove networks from others}
\usage{
exclude_networks(include, exclude)
}
\arguments{
\item{include}{An \code{\link{ip_network}} vector}

\item{exclude}{An \code{\link{ip_network}} vector}
}
\value{
An \code{\link{ip_network}} vector
}
\description{
Returns the minimal set of networks
}
