% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip_address.R
\name{ip_address}
\alias{ip_address}
\alias{is_ip_address}
\alias{format.vctrs_ip_address}
\title{IP address}
\usage{
ip_address(x = character())

is_ip_address(x)

\method{format}{vctrs_ip_address}(x, ...)
}
\arguments{
\item{x}{Vector of IP addresses}

\item{...}{Other parameters passed on}
}
\description{
IPv4 addresses are 32-bit integers that are usually presented as
a string in dotted-decimal notation, e.g. "192.168.0.1".

\code{ip_address()} is a high-level constructor that accepts character vectors
representing the IP address in dotted-decimal notation.

\code{is_ip_address()} checks if an object is of class \code{ip_address}.
}
