% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip_network.R
\name{ip_network}
\alias{ip_network}
\alias{is_ip_network}
\alias{format.vctrs_ip_network}
\title{IP network}
\usage{
ip_network(x = character())

is_ip_network(x)

\method{format}{vctrs_ip_network}(x, ...)
}
\arguments{
\item{x}{Vector of IP networks}

\item{...}{Other parameters passed on}
}
\description{
\code{ip_network()} is a high-level constructor that accepts character vectors
representing the IP network in CIDR notation.

\code{is_ip_network()} checks if an object is of class \code{ip_network}.
}
