% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipaddress-package.R
\docType{package}
\name{ipaddress-package}
\alias{ipaddress}
\alias{ipaddress-package}
\title{ipaddress: Tools for being tidy about IP addresses}
\description{
Implements an S3 class for storing IP addresses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://davidchall.github.io/ipaddress}
  \item \url{https://github.com/davidchall/ipaddress}
  \item Report bugs at \url{https://github.com/davidchall/ipaddress/issues}
}

}
\author{
\strong{Maintainer}: David Hall \email{david.hall.physics@gmail.com} (\href{https://orcid.org/0000-0002-2193-0480}{ORCID})

}
\keyword{internal}
