% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipaddress-package.R, R/operators.R
\name{ipaddress-vctrs}
\alias{ipaddress-vctrs}
\alias{vec_arith.ip_address}
\alias{vec_arith.ip_interface}
\title{Internal vctrs methods}
\usage{
\method{vec_arith}{ip_address}(op, x, y, ...)

\method{vec_arith}{ip_interface}(op, x, y, ...)
}
\description{
Internal vctrs methods
}
\keyword{internal}
