% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip_version.R
\name{is_ipv6}
\alias{is_ipv6}
\alias{is_ipv4}
\title{Version of the address space}
\usage{
is_ipv4(x)

is_ipv6(x)
}
\arguments{
\item{x}{An \code{\link{ip_address}} or \code{\link{ip_network}} vector}
}
\value{
A logical vector
}
\description{
Version of the address space
}
\examples{
ip <- ip_address(c("192.168.0.1", "2001:db8::7334"))

is_ipv4(ip)

is_ipv6(ip)
}
\seealso{
\code{\link[=max_prefix_length]{max_prefix_length()}}
}
