% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_mask.R
\name{netmask}
\alias{netmask}
\alias{hostmask}
\title{IP network masking}
\usage{
netmask(x)

hostmask(x)
}
\arguments{
\item{x}{An \code{\link{ip_network}} vector}
}
\value{
An \code{\link{ip_address}} vector.
}
\description{
These functions yield different representations of the IP network mask.
}
\examples{
x <- ip_network(c("192.168.0.0/22", "2001:db00::0/26"))

netmask(x)

hostmask(x)

}
\seealso{
The netmask can equivalently be represented by the \code{\link[=prefix_length]{prefix_length()}}.
}
