% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{netmask}
\alias{netmask}
\title{Extract netmask from IP network}
\usage{
netmask(network)
}
\arguments{
\item{network}{An \code{ip_network} vector}
}
\value{
An \code{ip_address} vector corresponding to the netmask
}
\description{
Extract netmask from IP network
}
\examples{
netmask(ip_network(c("192.168.1.0/24", "2001:db00::0/24")))
}
