% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip_network.R
\name{network_address}
\alias{network_address}
\alias{prefix_length}
\title{Extract components of an IP network}
\usage{
network_address(x)

prefix_length(x)
}
\arguments{
\item{x}{An \code{\link{ip_network}} vector}
}
\value{
\itemize{
\item \code{network_address()} returns an \code{\link{ip_address}} vector
\item \code{prefix_length()} returns an integer vector
}
}
\description{
An IP network is uniquely defined by its network address and
prefix length.
}
\examples{
x <- ip_network(c("192.168.0.0/22", "2001:db00::0/26"))

network_address(x)

prefix_length(x)

}
\seealso{
The prefix length can equivalently be represented by the \code{\link[=netmask]{netmask()}} or \code{\link[=hostmask]{hostmask()}}.
}
