% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample_ip}
\alias{sample_ip}
\alias{sample_ipv4}
\alias{sample_ipv6}
\title{Sample random addresses}
\usage{
sample_ipv4(size, replace = FALSE)

sample_ipv6(size, replace = FALSE)
}
\arguments{
\item{size}{Integer specifying the number of addresses to return}

\item{replace}{Should sampling be with replacement?}
}
\value{
An \code{\link{ip_address}} vector
}
\description{
Generates random IPv4 or IPv6 addresses.
}
\examples{
sample_ipv4(5)

sample_ipv6(5)
}
\seealso{
Use \code{\link[=sample_network]{sample_network()}} to sample addresses from a specific IP network.
}
